requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
black==24.3.0
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
Django==5.0.4
django-ninja==1.1.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
multidict==6.0.5
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
pydantic==2.6.4
pydantic_core==2.16.3
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
yarl==1.9.4
